refactor(routes): rename layout components for clarity
