+++ apps/web/src/AppRoutes.tsx
@@ -1,8 +1,8 @@
 import { FeatureFlagsKeysEnum } from '@novu/shared';
 import { Route, Routes } from 'react-router-dom';
-import { AppLayout } from './components/layout/AppLayout';
-import AuthLayout from './components/layout/components/AuthLayout';
-import { EnsureOnboardingComplete } from './components/layout/EnsureOnboardingComplete';
+import { PrivatePageLayout } from './components/layout/components/AppLayout';
+import { PublicPageLayout } from './components/layout/components/PublicPageLayout';
+import { EnsureOnboardingComplete } from './components/layout/components/EnsureOnboardingComplete';
 import { ROUTES } from './constants/routes.enum';
 import { useFeatureFlag } from './hooks';
 import { ActivitiesPage } from './pages/activities/ActivitiesPage';
@@ -54,7 +54,7 @@ export const AppRoutes = () => {
 
   return (
     <Routes>
-      <Route element={<AuthLayout />}>
+      <Route element={<PublicPageLayout />}>
         <Route path={ROUTES.AUTH_SIGNUP} element={<SignUpPage />} />
         <Route path={ROUTES.AUTH_LOGIN} element={<LoginPage />} />
         <Route path={ROUTES.AUTH_RESET_REQUEST} element={<PasswordResetPage />} />
@@ -78,7 +78,7 @@ export const AppRoutes = () => {
           }
         />
       </Route>
-      <Route element={<AppLayout />}>
+      <Route element={<PrivatePageLayout />}>
         <Route path={ROUTES.WORKFLOWS_DIGEST_PLAYGROUND} element={<TemplatesDigestPlaygroundPage />} />
         <Route path={ROUTES.WORKFLOWS_CREATE} element={<TemplateEditorPage />} />
         <Route path={ROUTES.WORKFLOWS_EDIT_TEMPLATEID} element={<TemplateEditorPage />}>
